--- .lycheeignore
@@ -1,3 +1,0 @@
-file:///*
-http://localhost:*
-http://notification-ws:8080
