--- .mdlrc
@@ -1,2 +1,0 @@
-rules "~MD002", "~MD003", "~MD005", "~MD007", "~MD013", "~MD022", "~MD024", "~MD027", "~MD028", "~MD029", "~MD033", "~MD034", "~MD036", "~MD041", "~MD055", "~MD057"
-style "#{File.dirname(__FILE__)}/markdownlint.rb"
