--- .yamllint
@@ -1,23 +1,0 @@
----
-extends: default
-rules:
-  line-length:
-    max: 256
-    level: warning
-  document-start: disable
-  braces:
-    forbid: false
-    min-spaces-inside: 0
-    max-spaces-inside: 1
-    min-spaces-inside-empty: -1
-    max-spaces-inside-empty: -1
-  commas:
-    max-spaces-before: -1
-    min-spaces-after: 1
-    max-spaces-after: -1
-  brackets:
-    max-spaces-inside: -1
-  empty-lines:
-    max: 3
-  indentation:
-    spaces: 2
