+++ Dockerfile
@@ -13,9 +13,9 @@
 # limitations under the License.
 FROM maven:3.8.3-openjdk-17 as build
 
-RUN mkdir /notification-ws
-COPY . /notification-ws
-RUN cd /notification-ws && mvn clean package -Dmaven.test.skip=true
+RUN mkdir /horten
+COPY . /horten
+RUN cd /horten && mvn clean package -Dmaven.test.skip=true
 
 FROM openjdk:17-jdk-alpine
 
@@ -24,6 +24,6 @@ LABEL maintainer-email="jack20220723@gmail.com"
 
 RUN apk add curl
 
-COPY --from=build /notification-ws/target/*.jar app.jar
+COPY --from=build /horten/target/*.jar app.jar
 
 ENTRYPOINT ["java","-jar","/app.jar"]
