+++ src/main/resources/application.properties
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-notification-ws.dingding.access-token=${NOTIFICATION_WS_DINGDING_ACCESS_TOKEN}
+horten.dingding.access-token=${HORTEN_DINGDING_ACCESS_TOKEN}
 
 server.forward-headers-strategy=framework
 springdoc.api-docs.enabled=true
