+++ src/test/java/org/qubitpi/horten/SwaggerIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.notification;
+package org.qubitpi.horten;
 
 import static org.assertj.core.api.Assertions.assertThat;
