+++ src/test/resources/application.properties
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-notification-ws.dingding.access-token=not-used
+horten.dingding.access-token=not-used
 
 server.forward-headers-strategy=framework
 springdoc.api-docs.enabled=true
